backend/requirements.txt
flask
flask-sqlalchemy
flask-mail
flask-socketio
flask-cors
mysqlclient
pyjwt
passlib
python-dotenv
requests